backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.1.2
google-genai==1.10.0
openai==1.58.1
anthropic==0.45.0
pydantic==2.10.4
python-dotenv==1.0.1
